requirements_conda.txt
requests
gdal=3.1.2
tabulate
bottle
PyYAML
Pillow
